Refactor display logging macros
