+++ src/cli/commands/install.rs
@@ -1,5 +1,3 @@
-use std::process::ExitCode;
-
 use clap::Args;
 
 use crate::{
