+++ src/cli/display/error.rs
@@ -1,0 +1,8 @@
+use thiserror::Error;
+
+/// The errors that occur during display.
+#[derive(Error, Debug)]
+pub enum DisplayError {
+    #[error("Could not display\nCaused by: {0}")]
+    IOError(#[from] std::io::Error),
+}
