+++ src/error_handling.rs
@@ -1,7 +1,6 @@
 use std::process::exit;
 
-use crate::cli::display;
-use crate::error_handling::display::display_error;
+use crate::cli::display::logging::error;
 
 pub trait HandleError<T> {
     fn unwrap_or_exit(self) -> T;
@@ -12,7 +11,7 @@ impl<T, E: std::fmt::Display> HandleError<T> for Result<T, E> {
         match self {
             Ok(value) => value,
             Err(e) => {
-                display_error(format_args!("{}", e));
+                error!("{}", e);
                 exit(1); // TODO: Make status code variable
             },
         }
