+++ src/repositories/manager.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use crate::{
-    cli::display,
+    cli::display::logging::warning,
     config::Config,
     installer::types::Version,
     platforms::TARGET_ARCHITECTURE,
@@ -26,7 +26,7 @@ impl<'a> RepositoryManager<'a> {
         for (id, repository) in &config.repositories {
             let provider = create_repository_provider(repository);
             if provider.is_none() {
-                display::warning!("Cannot create repository provider for repository {id}.");
+                warning!("Cannot create repository provider for repository {id}.");
                 continue;
             }
 
@@ -57,7 +57,7 @@ impl<'a> RepositoryManager<'a> {
             let provider = match self.providers.get(repository_id) {
                 Some(provider) => provider,
                 None => {
-                    display::warning!("Cannot find provider for {repository_id}, while it should exist.");
+                    warning!("Cannot find provider for {repository_id}, while it should exist.");
                     continue;
                 },
             };
@@ -106,7 +106,7 @@ impl<'a> RepositoryManager<'a> {
             let provider = match self.providers.get(repository_id) {
                 Some(provider) => provider,
                 None => {
-                    display::warning!("Cannot find provider for {repository_id}, while it should exist.");
+                    warning!("Cannot find provider for {repository_id}, while it should exist.");
                     continue;
                 },
             };
